✨ Modal para la eliminación de expedientes

+++ src/app/views/pages/expediente/listado-expediente/listado-expediente.component.ts
@@ -5,6 +5,9 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ModalExtenderSesionComponent } from 'src/app/views/components/shared/modal-extender-sesion/modal-extender-sesion.component';
+
 
 @Component({
   selector: 'app-listado-expediente',
@@ -19,7 +22,12 @@ export class ListadoExpedienteComponent implements OnInit {
   @ViewChild(MatSort) sort!: MatSort;
   visibleSpinner = false;
 
-  constructor( private pacienteService:DatosPersonalesService, private router:Router, private snack: MatSnackBar) { }
+  constructor( 
+    private pacienteService:DatosPersonalesService, 
+    private router:Router, 
+    private snack: MatSnackBar,
+    private dialog: MatDialog
+  ) { }
   
   ngOnInit(): void {
     this.cargarExpedientes();
@@ -48,32 +56,42 @@ export class ListadoExpedienteComponent implements OnInit {
   }
 
   eliminarExpediente(paciente: any){
-    this.visibleSpinner=true;
-    this.pacienteService.deletePaciente(paciente.id).subscribe({
-      next:(res)=>{
-        this.visibleSpinner=false;
-              this.snack.open(
-                res.mensaje,
-                'OK',
-                {
-                  duration: 3000,
-                }
-              );
-        this.cargarExpedientes();
-
-      },
-      error:(err)=>{
-        this.visibleSpinner=false;
-        this.snack.open(
-          err.mensaje,
-          'OK',
-          {
-            duration: 3000,
+    //Modal para solicitar confirmación de eliminación de expediente
+    const dialog = this.dialog.open( ModalExtenderSesionComponent, {
+      width: '30%',
+      data: {
+        titulo: 'Confirmar eliminación',
+        mensaje: '¿Desea eliminar de forma permanente el expediente seleccionado?',
+        boton: 'Aceptar'
+      }
+    });
+    dialog.afterClosed().subscribe(result => {
+      if(result) {
+        this.visibleSpinner=true;
+        this.pacienteService.deletePaciente(paciente.id).subscribe({
+          next:(res)=>{
+            this.visibleSpinner=false;
+                  this.snack.open(
+                    res.mensaje,
+                    'OK',
+                    {
+                      duration: 3000,
+                    }
+                  );
+            this.cargarExpedientes();
+          },
+          error:(err)=>{
+            this.visibleSpinner=false;
+            this.snack.open(
+              err.mensaje,
+              'OK',
+              {
+                duration: 3000,
+              }
+            );     
           }
-        );
-              
+        });
       }
     });
   }
-
 }
